+++ R/utils.R
@@ -121,10 +121,9 @@ parse_units <- function(unit) {
   }
 }
 
-
 # Because `as.POSIXct.Date()` always uses local timezone
 date2posixct <- function(x) {
-  out <- unclass(x) * 86400
-  attributes(out) <- list(tzone = "UTC", class = c("POSIXct", "POSIXt"))
-  out
+  structure(unclass(x) * 86400,
+            tzone = "UTC",
+            class = c("POSIXct", "POSIXt"))
 }