+++ NEWS.md
@@ -5,6 +5,7 @@ Version 0.1.1
 
  * Follow vctrs replication rules
  * Change arguments of `time_add()` and `time_subtract()` to singulars
+ * Build on top of cpp11 instead of Rcpp
 
 Version 0.1.0
 =============